refactor(lab): extract BottomSheet keyboard hook
